requirements.txt
pandas>=1.3.0
pyarrow>=6.0.0
pyspark>=3.0.0
python-dotenv>=0.19.0